fix bugs generated in large data volumes

+++ src/femu/nvme.h
@@ -33,7 +33,7 @@ typedef struct NvmeCmd {
 typedef struct NvmeRequest {
     uint64_t                id;
     uint64_t                slba;
-    uint16_t                nlb;
+    uint64_t                nlb;
     NvmeCmd                 cmd;
     int64_t                 stime;
     int64_t                 reqlat;
