+++ src/include/femu.h
@@ -11,10 +11,10 @@ extern "C"{
 
 struct FemuCtrl;
 
-struct FemuCtrl *femu_init(size_t ssd_size, bool enable_backend, bool enable_latency);
-int femu_read(struct FemuCtrl *femu, size_t len, off_t offset, void *arg);
-int femu_write(struct FemuCtrl *femu, size_t len, off_t offset, void *arg);
-int femu_reset(struct FemuCtrl *femu, size_t len, off_t offset, void *arg);
+struct FemuCtrl *femu_init(uint64_t ssd_size, bool enable_backend, bool enable_latency);
+int femu_read(struct FemuCtrl *femu, uint64_t len, uint64_t offset, void *arg);
+int femu_write(struct FemuCtrl *femu, uint64_t len, uint64_t offset, void *arg);
+int femu_reset(struct FemuCtrl *femu, uint64_t len, uint64_t offset, void *arg);
 void get_zns_meta(uint64_t *meta);
 
 #ifdef __cplusplus
